WSWAV-17 Week 1 Tasks
Create the base structure of the black box under “Scan Results”, the base structure of found issues in the “Findings” Section, and create the base structure of the “Scan Coverage Information” Section.

The Scan Results page must be fully responsive to accommodate screens that are a minimum of 280 pixels wide.